Store open windows for non-standard slots
Update the save file to change how OpenSlots are tracked. Before, StringHash would be zero if the item was not a standard slot (e.g. belt, jetpack, hand, etc).

We will try to use the ReferenceId in the StringHash field when the StringHash is zero. On load, we will try to find a matching slot for that ReferenceId so that we can re-open that window.

Since this will modify a user's save, add a configuration option which defaults to false. While the current game version (0.2.3920.18842) has no issues and ignores invalid StringHash values, this may not be the case in the future.